+++ docs/CHANGELOG.md
@@ -1,8 +1,19 @@
 # Releases
 
-## v1.42.0
+## v1.43.0
 
 Released: ???
+
+Features:
+- Added custom Fishing Hook timer which extends functionality of default Hypixel's one. Explore /feesh -> Rendering -> Fishing Hook section to enable and customize.
+- Added option to set a keybind in Minecraft's Controls menu to share the nearest Hotspot to PARTY or ALL chat on button pressed.
+
+Bugfixes:
+-
+
+## v1.42.0
+
+Released: 2025-05-20
 
 Features:
 - Sea creature HP tracker: Added immunity indicator for sea creatures that have 5 seconds damage reduction period.